实体类更新

+++ xiu-service/src/main/java/com/vick/xiu/entity/Score.java
@@ -1,11 +1,12 @@
 package com.vick.xiu.entity;
 
-import java.io.Serializable;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -13,7 +14,7 @@ import lombok.experimental.Accessors;
  * </p>
  *
  * @author zyz
- * @since 2019-11-10
+ * @since 2019-11-25
  */
 @Data
 @EqualsAndHashCode(callSuper = false)
@@ -24,13 +25,13 @@ public class Score implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @ApiModelProperty(value = "用户ID")
-    private Integer userId;
+    private Long userId;
 
     @ApiModelProperty(value = "课程ID")
-    private Integer courseId;
+    private Long courseId;
 
     @ApiModelProperty(value = "考试ID")
-    private Integer examId;
+    private Long examId;
 
     @ApiModelProperty(value = "分数")
     private Float score;
